+++ sacred/randomness/__init__.py
@@ -1,0 +1,5 @@
+import sacred
+
+
+def set_python_random_seed(seed):
+    sacred.randomness.python.set_seed(seed)
