+++ sacred/randomness/tensorflow.py
@@ -1,0 +1,6 @@
+from sacred import optional as opt
+
+
+def set_seed(seed):
+    tf = opt.get_tensorflow()
+    tf.set_random_seed(seed)
