+++ sacred/run.py
@@ -10,7 +10,7 @@ import warnings
 import sacred
 from sacred import metrics_logger
 from sacred.metrics_logger import linearize_metrics
-from sacred.randomness import set_global_seed
+from sacred.randomness.legacy import set_global_seed
 from sacred.utils import SacredInterrupt, join_paths, IntervalTimer
 from sacred.stdout_capturing import get_stdcapturer
